second upload

+++ 0x05-pointers_arrays_strings/1-swap.c
@@ -8,7 +8,7 @@
  */
 void swap_int(int *a, int *b)
 {
-	int temp;
+	long int temp;
 
 	temp = *a;
 	*a = *b;
